requirements.txt
# Core dependencies
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0

# Data analysis and AI
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
scikit-learn==1.3.2

# Web framework for health checks
flask==3.0.0

# WebSocket for real-time data
websocket-client==1.7.0

# Scheduling
schedule==1.2.0
